Removes ethereumjs-util and keccak dependencies.
Replaces both with `js-sha3`.  `ethereumjs-util` and `keccak` NPM packages both have native bindings in their dependency tree which means they depend on node-gyp along the way.  `js-sha3` on the other hand is a native implementation of the sha3 family of hashing functions, which relieves a number of `node-gyp` and native binding related issues.